Overhaul of AutoFill + Features
Edited some auto fill parameters.
Autofill is now based on number of points for now, until we get conditions working.
Reduce autofill quality, for app speed up.
Enabled cleanup of a layer if it is going to be auto filled, works most of the time.
Bug catching, and fixes.
Menu Corrections and fixes.
Update Saving feature.